Pulled RandomGen and related functions into RageUtil_RandomNumbers file
