+++ src/BGAnimationLayer.cpp
@@ -13,6 +13,7 @@
 #include "LuaManager.h"
 #include "AutoActor.h"
 #include "ThemeManager.h"
+#include "RageUtil_RandomNumbers.h"
 
 #include <cmath>
