+++ src/BPMDisplay.cpp
@@ -10,6 +10,7 @@
 #include "LocalizedString.h"
 #include "Song.h"
 #include "Steps.h"
+#include "RageUtil_RandomNumbers.h"
 
 #include <cmath>
 #include <climits>
