+++ src/CMakeData-rage.cmake
@@ -6,6 +6,7 @@ list(APPEND SMDATA_RAGE_UTILS_SRC
             "RageUtil_CharConversions.cpp"
             "RageUtil_FileDB.cpp"
             "RageUtil_LanguageInfo.cpp"
+            "RageUtil_RandomNumbers.cpp"
             "RageUtil_WorkerThread.cpp")
 
 list(APPEND SMDATA_RAGE_UTILS_HPP
@@ -17,6 +18,7 @@ list(APPEND SMDATA_RAGE_UTILS_HPP
             "RageUtil_CircularBuffer.h"
             "RageUtil_FileDB.h"
             "RageUtil_LanguageInfo.h"
+            "RageUtil_RandomNumbers.h"
             "RageUtil_WorkerThread.h")
 
 source_group("Rage\\\\Utils"
