+++ src/Character.cpp
@@ -4,6 +4,7 @@
 #include "RageUtil.h"
 #include "RageTextureID.h"
 #include "ActorUtil.h"
+#include "RageUtil_RandomNumbers.h"
 
 #include <vector>
