+++ src/CryptManager.cpp
@@ -15,6 +15,7 @@
 #include "LuaBinding.h"
 #include "LuaReference.h"
 #include "LuaManager.h"
+#include "RageUtil_RandomNumbers.h"
 
 #include <cstdint>
 #include <vector>
