+++ src/Inventory.cpp
@@ -10,6 +10,7 @@
 #include "StatsManager.h"
 #include "ThemeMetric.h"
 #include "PlayerState.h"
+#include "RageUtil_RandomNumbers.h"
 
 #include <vector>
