+++ src/PlayerAI.cpp
@@ -4,6 +4,7 @@
 #include "IniFile.h"
 #include "GameState.h"
 #include "PlayerState.h"
+#include "RageUtil_RandomNumbers.h"
 
 #define AI_PATH "Data/AI.ini"
