+++ src/ScoreKeeperRave.cpp
@@ -9,6 +9,7 @@
 #include "ThemeMetric.h"
 #include "PlayerState.h"
 #include "NoteTypes.h"
+#include "RageUtil_RandomNumbers.h"
 
 #include <cstddef>
