+++ src/ScreenOptionsCourseOverview.cpp
@@ -17,6 +17,7 @@
 #include "PlayerState.h"
 #include "Style.h"
 #include "PrefsManager.h"
+#include "RageUtil_RandomNumbers.h"
 
 #include <vector>
