+++ src/arch/InputHandler/InputHandler_MonkeyKeyboard.cpp
@@ -2,6 +2,7 @@
 #include "InputHandler_MonkeyKeyboard.h"
 #include "RageUtil.h"
 #include "PrefsManager.h"
+#include "RageUtil_RandomNumbers.h"
 
 #include <vector>
